requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-environ==0.8.1
django-mptt==0.13.4
django-mptt-admin==2.2.0
djangorestframework==3.13.1
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
yarl==1.7.2
python-dotenv~=0.20.0
